requirements.txt
xlrd==2.0.1                    # Чтение старых Excel файлов (.xls)
sqlalchemy==2.0.23             # ORM для более удобной работы с БД (альтернатива)
mysql-connector-python==8.2.0  # Для подключения к MySQL
pandas==2.1.4                  # Работа с данными (CSV, Excel)
openpyxl==3.1.2                # Чтение Excel файлов (.xlsx)
rich==13.7.0                   # Для красивого отображения информации в консоли
tqdm==4.66.1                   # Прогресс-бары для длительных операций
python-dotenv==1.1.0           # Работа с переменными окружения